fs: add fs_readdir()
Needed to support efi file protocol.  The fallback.efi loader wants
to be able to read the contents of the /EFI directory to find an OS
to boot.

For reference, the expected EFI semantics are described in (v2.7 of UEFI
spec) in section 13.5 (page 609).  Or for convenience, see:

  http://wiki.phoenix.com/wiki/index.php/EFI_FILE_PROTOCOL#Read.28.29

The EFI level semantics are implemented in a later patch, so they are
not too important to the understanding of this patch.

Signed-off-by: Rob Clark <robdclark@gmail.com>